requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
biopython==1.84
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
h5py==3.12.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.11.9
logomaker==0.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchmetrics==1.6.0
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
wheel==0.45.1
yarl==1.18.3
